feat: redis 이용한 장바구니 구현

+++ src/main/java/com/nhnacademy/bookstorefront/bookcart/dto/response/GetBookCartResponse.java
@@ -16,6 +16,6 @@ public record GetBookCartResponse(
 	BigDecimal bookSalePrice,
 	BigDecimal bookSalePercent,
 	int inventorQuantity,
-	int bookQuantity
+	Long bookQuantity
 ) {
 }
